Package rename from suddenpluscalculator to iidxtools
