+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdkVersion 26
     buildToolsVersion '26.0.2'
     defaultConfig {
-        applicationId "com.duraz0rz.suddenpluscalculator"
+        applicationId "com.duraz0rz.iidxtools"
         minSdkVersion 26
         targetSdkVersion 26
         versionCode 1
