+++ app/src/androidTest/java/com/duraz0rz/iidxtools/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.duraz0rz.suddenpluscalculator
+package com.duraz0rz.iidxtools
 
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("com.duraz0rz.suddenpluscalculator", appContext.packageName)
+        assertEquals("com.duraz0rz.iidxtools", appContext.packageName)
     }
 }
