+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.duraz0rz.suddenpluscalculator">
+    package="com.duraz0rz.iidxtools">
 
     <application
         android:allowBackup="true"
@@ -9,13 +9,13 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activities.SuddenPlusInputActivity">
+        <activity android:name="com.duraz0rz.iidxtools.activities.SuddenPlusInputActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activities.SuddenPlusTableActivity" android:parentActivityName=".activities.SuddenPlusInputActivity" />
+        <activity android:name="com.duraz0rz.iidxtools.activities.SuddenPlusTableActivity" android:parentActivityName="com.duraz0rz.iidxtools.activities.SuddenPlusInputActivity" />
     </application>
 
 </manifest>
