+++ app/src/main/java/com/duraz0rz/iidxtools/dataClasses/DataClasses.kt
@@ -1,3 +1,3 @@
-package com.duraz0rz.suddenpluscalculator.dataClasses
+package com.duraz0rz.iidxtools.dataClasses
 
 data class SuddenPlusValue(val highSpeed : String, val minWhiteNumber: Long, val maxWhiteNumber: Long? = null)
